Add shared SQL reader and writer bases
