+++ src/etl_core/components/databases/sqlserver/sqlserver_write.py
@@ -1,25 +1,21 @@
 from __future__ import annotations
 
-from typing import Any, AsyncIterator, Dict, Optional  # noqa: F401
-
-import dask.dataframe as dd
-import pandas as pd
+from typing import Optional  # noqa: F401
 from pydantic import model_validator
 
 from etl_core.components.databases.sqlserver.sqlserver import SQLServerComponent
 from etl_core.components.databases.database_operation_mixin import (
     DatabaseOperationMixin,
 )
+from etl_core.components.databases.sql_writer_base import SQLWriterBase
 from etl_core.components.databases.if_exists_strategy import DatabaseOperation
 from etl_core.components.component_registry import register_component
-from etl_core.metrics.component_metrics.component_metrics import ComponentMetrics
 from etl_core.receivers.databases.sqlserver.sqlserver_receiver import SQLServerReceiver
-from etl_core.components.envelopes import Out
 from etl_core.components.wiring.ports import InPortSpec, OutPortSpec
 
 
 @register_component("write_sqlserver")
-class SQLServerWrite(SQLServerComponent, DatabaseOperationMixin):
+class SQLServerWrite(SQLWriterBase, SQLServerComponent, DatabaseOperationMixin):
     """
     SQL Server writer with ports + schema.
 
@@ -31,6 +27,8 @@ class SQLServerWrite(SQLServerComponent, DatabaseOperationMixin):
 
     INPUT_PORTS = (InPortSpec(name="in", required=True, fanin="many"),)
     OUTPUT_PORTS = (OutPortSpec(name="out", required=False, fanout="many"),)
+
+    receiver_class = SQLServerReceiver
 
     def _build_query(
         self, table: str, columns: list, operation: DatabaseOperation, **kwargs
@@ -94,50 +92,11 @@ class SQLServerWrite(SQLServerComponent, DatabaseOperationMixin):
     @model_validator(mode="after")
     def _build_objects(self):
         """Build SQL Server-specific objects after validation."""
-        self._receiver = SQLServerReceiver()
+        self._initialize_receiver()
         schema = self.in_port_schemas["in"]
         columns = [field.name for field in schema.fields]
         self._query = self._build_query(self.entity_name, columns, self.operation)
         return self
-
-    async def process_row(
-        self, row: Dict[str, Any], metrics: ComponentMetrics
-    ) -> AsyncIterator[Out]:
-        """Write a single row and emit it (or receiver result) on 'out'."""
-        result = await self._receiver.write_row(
-            entity_name=self.entity_name,
-            row=row,
-            metrics=metrics,
-            query=self._query,
-            connection_handler=self.connection_handler,
-        )
-        yield Out(port="out", payload=result)
-
-    async def process_bulk(
-        self, data: pd.DataFrame, metrics: ComponentMetrics
-    ) -> AsyncIterator[Out]:
-        """Write a pandas DataFrame and emit the same frame on 'out'."""
-        result = await self._receiver.write_bulk(
-            entity_name=self.entity_name,
-            frame=data,
-            metrics=metrics,
-            query=self._query,
-            connection_handler=self.connection_handler,
-        )
-        yield Out(port="out", payload=result)
-
-    async def process_bigdata(
-        self, ddf: dd.DataFrame, metrics: ComponentMetrics
-    ) -> AsyncIterator[Out]:
-        """Write a Dask DataFrame and emit the same ddf on 'out'."""
-        result = await self._receiver.write_bigdata(
-            entity_name=self.entity_name,
-            frame=ddf,
-            metrics=metrics,
-            query=self._query,
-            connection_handler=self.connection_handler,
-        )
-        yield Out(port="out", payload=result)
 
 
 SQLServerWrite.model_rebuild()
